+++ client/components/Procom.js
@@ -2,6 +2,7 @@ import React from 'react';
 import { connect } from 'react-redux';
 import { createMarkup } from '../helpers';
 import PasoOptions from './PasoOptions';
+import TextArea from './TextArea';
 
 class Procom extends React.Component {
   constructor(props) {
@@ -14,27 +15,18 @@ class Procom extends React.Component {
     }
 
     this.procomFormSubmiter = this.procomFormSubmiter.bind(this);
-    this.setTextareaHeight = this.setTextareaHeight.bind(this);
     this.onChange4Textarea = this.onChange4Textarea.bind(this);
     this.procomOptionsConection = this.procomOptionsConection.bind(this);
   }
 
   componentDidMount(){
-    this.setTextareaHeight($('textarea'));
     let estilo = this.props.procom.pro_style;
     this.setState({estilo})
-  }
-
-  setTextareaHeight(paso){
-    paso.each(function(index, item){
-      item.style.height = item.scrollHeight+'px';
-    });
   }
 
   // EDIT PROCOM
   onChange4Textarea(viewType){
     this.setState({showEditButtons: viewType});
-    this.setTextareaHeight($('#edit-procom-textarea'));
   }
 
   // procom connection with PasoOptions component
@@ -55,7 +47,7 @@ class Procom extends React.Component {
     let pasId = this.props.pasoId;
     let proId = this.props.proyectoId;
     let procom = this.props.procom;
-    let pro_content = this.refs.pro_content.value;
+    let pro_content = this.refs.text_area.refs.text_content.value;
     let type_of_issue;
     let pro_style;
     if(this.state.estilo === 'problema') {
@@ -90,15 +82,13 @@ class Procom extends React.Component {
   render(){
     // <strong>{`${this.props.numeracion} `}</strong>
     let whichButtonsShouldIHave = 'add-procom-full-buttons'
-
     let procom = this.props.procom;
     let pasoId = this.props.pasoId;
     let proyectoId = this.props.proyectoId;
     let comentario = true;
     let problema = false;
-    let show = 'show-buttons';
     let hide = 'hide-buttons';
-    let inlineStyle = {height: '18px'};
+
     return (
       <div className={`procom-container ${procom.pro_style}`}>
         {this.extraContent()}
@@ -106,7 +96,7 @@ class Procom extends React.Component {
           <div className={this.state.showEditButtons}>
             <PasoOptions whichType={whichButtonsShouldIHave} elected={this.state.estilo} conection={this.procomOptionsConection} />
           </div>
-          <textarea id="edit-procom-textarea" className="procom-content-text" style={inlineStyle} ref='pro_content' onChange={()=>this.onChange4Textarea(show)} defaultValue={procom.pro_content}></textarea>
+          <TextArea the_class='procom-content-text' the_content={procom.pro_content} onChange4Textarea={this.onChange4Textarea} ref='text_area'/>
         </span>
         <span className="mini-botones">
           <i className="fa fa-trash" aria-hidden="true" onClick={()=> this.props.deleteProcomFunc(procom.id, pasoId, proyectoId)}></i>
